requirements.txt
fastapi
uvicorn[standard]
supabase
qdrant-client>=1.16.0
python-dotenv
pydantic
numpy
openai
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
transformers>=4.30.0
scikit-learn>=1.3.0
pint
quantulum3
requests
nltk
